Updates
I can't figure out how to change the variable "attempts", and carry it over to the completion page for the levels.

The help button is problematic. I can click on it and have it show up, and then I need to remove it, so it returns to the current level.

I would like to figure out how to go to create a "levels" page, but it's not that big of a deal. I don't think I have the time to implement it.